--- code/file_uplaod_app/.idea/.gitignore
@@ -1,2 +1,0 @@
-# Default ignored files
-/workspace.xml
