--- code/file_uplaod_app/src/commonComponents/uploadDropbox.css
@@ -1,34 +1,0 @@
-@import '~antd/dist/antd.css';
-
-.fileDropBox {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    border: 1px dashed rgb(212,212,212);
-    background-color: rgb(255,255,255);
-    color: rgb(137,137,137);
-    text-align: center;
-}
-.fileDropBox:hover {
-    background-color: rgb(248,248,248);
-    border: 1px dashed rgb(55,159,255);
-}
-.uploadDragIcon {
-    display: block;
-    margin-top: 100px;
-    font-size: 120px;
-    color: rgb(55,159,255);
-}
-.uploadTips {
-    display: block;
-    font-size: 18px;
-}
-.fileDropBox>input {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    top: 0;
-    left: 0;
-    opacity: 0;
-}
-
