--- code/file_uplaod_app/src/index.js
@@ -1,9 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import APP from "./App";
-
-
-ReactDOM.render(
-    <APP />,
-    document.getElementById("root")
-);
