--- code/file_uplaod_app/src/jsPages/fileExtract/extractPage.css
@@ -1,59 +1,0 @@
-* {
-    padding: 0;
-    margin: 0;
-}
-.extractPageComponent {
-    width: 100%;
-    height: 100%;
-}
-.extractLeft {
-    float: left;
-    width: 20%;
-    height: 100%;
-    border: 1px solid rgb(238,240,243);
-}
-.listTitle {
-    font-size: 25px;
-    padding-left: 10px;
-    border-bottom: 1px solid rgb(238,240,243);
-}
-.extractRight {
-    float: left;
-    width: 75%;
-    height: 100%;
-    margin-left: 10px;
-}
-.extractTop {
-    width: 100%;
-    height: 25%;
-}
-#fileContent {
-    width: 100%;
-    height: 100%;
-    resize: none;
-    padding-left: 20px;
-    font-size: 18px;
-    overflow: scroll;
-    border: 1px solid rgb(238,240,243);
-}
-#fileContent::-webkit-scrollbar{
-    display: none;
-}
-.extractBottom {
-    width: 100%;
-    height: 75%;
-    margin-top: 10px;
-    overflow: scroll;
-}
-.extractBottom::-webkit-scrollbar {
-    display: none;
-}
-.infoFormBox {
-    width: 100%;
-    height: 100%;
-    padding-right: 100px;
-    margin: 0 auto;
-}
-.ant-btn {
-    margin-right: 8px;
-}
