--- code/file_uplaod_app/src/jsPages/fileUpload/uploadPage.css
@@ -1,4 +1,0 @@
-* {
-    padding: 0;
-    margin: 0;
-}
