+++ code/file_upload/package.json
@@ -5,7 +5,7 @@
   },
   "version": "0.1.0",
   "private": true,
-  "homepage": "https://darrenqiaoo.github.io/summercamp-p5", 
+  "homepage": "https://darrenqiaoo.github.io/summercamp-p5",
   "dependencies": {
     "@ant-design/icons": "^4.2.1",
     "@testing-library/jest-dom": "^4.2.4",
