+++ code/file_upload/src/App.js
@@ -1,15 +1,12 @@
-import React from 'react';
-import './App.css';
-import {HashRouter as Router, Route} from 'react-router-dom';
+import React from "react";
+import {BrowserRouter as Router, Route} from "react-router-dom";
 import MyTemplate from "./jsPages/template/template";
-//import Test from "./tests/test";
 
 function App() {
     return (
         <Router>
             <div>
                 <Route exact path={"/"} component={MyTemplate}/>
-                {/*<Route path={'/test'} component={Test}/>*/}
             </div>
         </Router>
     );
