+++ code/file_upload/src/commonComponents/fileList.js
@@ -1,26 +1,26 @@
 import React from "react";
-import { DeleteOutlined, PaperClipOutlined } from '@ant-design/icons';
+import { DeleteOutlined, PaperClipOutlined } from "@ant-design/icons";
 import { message } from "antd";
 import "./fileList.css"
 
 class FileList extends React.Component{
     // 文件列表组件
-    itemClick(file_name){
+    itemClick(fileName){
         // 列表元素点击事件
-        this.props.showFileContent(file_name);
+        this.props.showFileContent(fileName);
     };
-    handleDelete(file_name){
+    handleDelete(fileName){
         // 删除图标点击事件
         let tempFiles = []; // 记录不删除的文件名
         for(let i=0;i<this.props.files.length;i++){
-            if(this.props.files[i]!==file_name){
+            if(this.props.files[i]!==fileName){
                 // 文件名不是当前眼删除的文件名，记录到tempFiles
                 tempFiles.push(this.props.files[i]);
             }
         }
         this.props.setFiles(tempFiles); // 向父组件返回删除后的文件名列表
-        localStorage.removeItem(file_name); // 删除localStorage中文件名对应的记录
-        message.success(file_name+" has been deleted.");
+        localStorage.removeItem(fileName); // 删除localStorage中文件名对应的记录
+        message.success(fileName+" has been deleted.");
     }
     render() {
         const fList = this.props.files.map(f => {
