--- code/file_upload/src/commonComponents/fileShowBox.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-class FileShowBox extends React.Component{
-    render() {
-        return(
-            <div className={this.props.className}>
-                <textarea id={this.props.id} value={this.props.value} readOnly={true}/>
-            </div>
-        );
-    }
-}
-
-export default FileShowBox;
