+++ code/file_upload/src/index.js
@@ -1,8 +1,9 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import React from "react";
+import ReactDOM from "react-dom";
 import APP from "./App";
+
 
 ReactDOM.render(
     <APP />,
-    document.getElementById('root')
+    document.getElementById("root")
 );
