+++ code/file_upload/src/jsPages/fileExtract/extractPage.css
@@ -10,7 +10,7 @@
     float: left;
     width: 20%;
     height: 100%;
-    border: 1px solid rgb(238,240,243);;
+    border: 1px solid rgb(238,240,243);
 }
 .listTitle {
     font-size: 25px;
