+++ code/file_upload/src/jsPages/fileUpload/uploadPage.css
@@ -1,0 +1,4 @@
+* {
+    padding: 0;
+    margin: 0;
+}
